+ Added comparison operator for Vertex + Model objects can now be constructed from obj::Object objects - Changed raw arrays for std::array - Renamed member data to keep with style guide - Removed Count from enums - Removed warnings

+++ src/GLMatrix.cpp
@@ -13,7 +13,7 @@ namespace glt
     {
         for ( int index = 0; index < 16; ++index )
         {
-            m_data[ index ] = (float) i_matrix[ index ];
+            m_data[ (long long unsigned int)index ] = (float) i_matrix[ index ];
         }
     }
 }
